requirements.txt
suds-py3==1.4.4.1
spyne==2.13.16
pymongo==3.11.2